@@ -10,17 +10,14 @@ Gem::Specification.new do |s|
 
   s.required_rubygems_version = ">= 1.3.6"
   s.platform = Gem::Platform::RUBY
-  s.add_dependency('rails', '>= 3.1')
-  s.add_dependency('prawn', '>= 1.0.0.rc2')
-  # s.add_dependency('prawn', '>= 0.12.0')
+  s.add_dependency('rails', '>= 3.2')
+  s.add_dependency('prawn', '>= 0.13')
 
-  s.add_development_dependency "appraisal"
-  s.add_development_dependency "rspec-rails", "2.13.2"
+  s.add_development_dependency "rspec-rails", "2.14"
   s.add_development_dependency "mocha", "0.14.0"
 
-  s.files         = `git ls-files`.split("\n")
-  s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
-  s.require_path  = "lib"
+  s.files = Dir["{app,config,db,lib}/**/*"] + ["MIT-LICENSE", "Rakefile", "README.rdoc"]
+  s.test_files = Dir["test/**/*"]
 
   s.has_rdoc = true
   s.extra_rdoc_files = ["README.rdoc"]
