@@ -1,58 +1,62 @@
 require "spec_helper"
 
-describe TestController, :type => :controller do
+describe TestController do
+  render_views
 
   describe "simple" do
     it "returns correct PDF" do
-      get "/default_render.pdf"
-      response.should be_success
+      get :default_render, :format => :pdf
+      expect(response).to be_success
 
       asset_binary = File.open(asset_path("default_render")).read.bytes.to_a
       body_binary = response.body.bytes.to_a
-      body_binary.should == asset_binary
+      expect(body_binary).to eq(asset_binary)
     end
 
 
     it "shares values/changes of instance vars between view and helpers" do
-      expect { get "/instance_var_test.pdf" }.to_not raise_error
+      expect { get :instance_var_test, :format => :pdf }.to_not raise_error
     end
 
     it "should render items in a block passed to a helper" do
-      get "/yield_block_in_helper_test.pdf"
+      get :yield_block_in_helper_test, :format => :pdf
+      expect(response).to be_success
+
       asset_binary = File.open(asset_path("yield_block_in_helper_test")).read.bytes.to_a
       body_binary = response.body.bytes.to_a
-      body_binary.should == asset_binary
+      expect(body_binary).to eq(asset_binary)
     end
   end
 
 
   describe "dsl" do
     it "returns correct PDF" do
-      get "/dsl_render.pdf"
-      response.should be_success
+      get :dsl_render, :format => :pdf
+      expect(response).to be_success
 
       asset_binary = File.open(asset_path("dsl_render")).read.bytes.to_a
       body_binary = response.body.bytes.to_a
-      body_binary.should == asset_binary
+      expect(body_binary).to eq(asset_binary)
     end
   end
 
   describe "partials" do
     it "renders partials" do
-      get "/partial_render.pdf"
-      response.should be_success
+      get :partial_render, :format => :pdf
+      expect(response).to be_success
 
       asset_binary = File.open(asset_path("partial_render")).read.bytes.to_a
       body_binary = response.body.bytes.to_a
-      body_binary.should == asset_binary
+      expect(body_binary).to eq(asset_binary)
     end
   end
 
   describe "complex headers" do
     it "should return a file with a specified filename" do
-      get "/filename_test.pdf"
-      response.should be_success
-      response.header["Content-Disposition"].should == "attachment;filename=\"fancy_name.pdf\""
+      get :filename_test, :format => :pdf
+      expect(response).to be_success
+
+      expect(response.header["Content-Disposition"]).to eq("attachment;filename=\"fancy_name.pdf\"")
     end
 
   end
